Move Mcap0IndexedReader out of subfolder
