+++ typescript/core/src/index.ts
@@ -2,7 +2,7 @@ export { default as McapPre0Reader } from "./pre0/McapPre0Reader";
 export { default as McapPre0Writer } from "./pre0/McapPre0Writer";
 export * as McapPre0Types from "./pre0/types";
 
-export { default as Mcap0IndexedReader } from "./v0/Mcap0IndexedReader";
+export { Mcap0IndexedReader } from "./v0/Mcap0IndexedReader";
 export { default as Mcap0StreamReader } from "./v0/Mcap0StreamReader";
 export { Mcap0Writer } from "./v0/Mcap0Writer";
 export type { Mcap0WriterOptions } from "./v0/Mcap0Writer";
