+++ typescript/core/src/v0/Mcap0IndexedReader.ts
@@ -1,12 +1,12 @@
 import { crc32, crc32Final, crc32Init, crc32Update } from "@foxglove/crc";
 import Heap from "heap-js";
 
-import { MCAP0_MAGIC } from "../constants";
-import { parseMagic, parseRecord } from "../parse";
-import { DecompressHandlers, IReadable, TypedMcapRecords } from "../types";
+import { MCAP0_MAGIC } from "./constants";
+import { parseMagic, parseRecord } from "./parse";
+import { DecompressHandlers, IReadable, TypedMcapRecords } from "./types";
 import { ChunkCursor } from "./ChunkCursor";
 
-export default class Mcap0IndexedReader {
+export class Mcap0IndexedReader {
   readonly chunkIndexes: readonly TypedMcapRecords["ChunkIndex"][];
   readonly attachmentIndexes: readonly TypedMcapRecords["AttachmentIndex"][];
   readonly metadataIndexes: readonly TypedMcapRecords["MetadataIndex"][] = [];
